requirements.txt
# Django and REST Framework
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.10

# JWT Authentication and Email
djangorestframework-simplejwt==5.3.0
django-templated-mail==1.1.1

# Cloudinary for image storage
cloudinary==1.41.0

# Environment variables
python-decouple==3.8
python-dotenv==1.0.1

# API Documentation
drf-spectacular==0.27.2

# Additional utilities
Pillow==10.4.0
python-dateutil==2.9.0
pytz==2024.2

# Production server
gunicorn==21.2.0

# Development dependencies
django-debug-toolbar==4.4.6

# Testing dependencies
factory-boy==3.3.1
coverage==7.6.1
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
pytest-mock==3.14.0
model-bakery==1.19.5

# AWS dependencies
boto3==1.34.34
django-storages==1.14.2
django-ses==3.5.2

# Monitoring and error tracking
sentry-sdk==1.40.3

# Caching
django-redis==5.4.0

# Security
django-ratelimit==4.1.0